# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 学习状态智能助手 API（板块一 + 板块二 MVP）
  version: 1.8.0
  description: |
    中学学情导学与状态评估 API。

    来源：`docs/api-design-unified.md`（唯一生效版本），对应 `PRD-学习状态智能助手.md`。
    本文件严格按该设计文档生成，未引入文档之外的字段。

    主流程：创建目标 ① → 生成计划 ② → 提交学习记录 ③ → 计算状态 ④ → 返回建议 ⑤（外加周期复盘与反馈回路）。

    约定要点：
    - 所有资源以当前 token 用户为作用域，不设 userId 路径参数，不存在跨用户读取接口。
    - 写接口可选携带 `Idempotency-Key` 头，24 小时内重复键直接返回首次结果。
    - 建议与复盘为异步生成：POST 返回 202 + `generation.status = pending`，前端轮询 GET 获取结果。
    - 监护人授权失效时所有写接口返回 403 `GUARDIAN_AUTHORIZATION_EXPIRED`，读接口不受影响。

    阶段口径：pilot。本文档为前后端与 QA 的唯一契约来源。项目当前处于 pilot 阶段（小范围试用验证，非 beta/正式发布），由学生团队维持。板块三契约已于 M1 转正（口径见 docs/module3-m0-decision-proposal.md v1.7 §4.11）。

    **v1.6.0（重构 M0 · 契约冻结）**——本次为重构第一阶段的契约增量，未改动任何既有字段名：
    - **稳定用户 ID 落地（D59）**：`User.userId` 语义明确为稳定内部主键（`u_` 前缀），与登录凭证
      解耦；`AuthMeResponse.user` 增补可选 `userId`（`email` 保留，前端展示仍用它）。
    - **新增 M0 实体 schema**：Exam / Collection / CollectionItem / UsageLedgerEntry / InviteCode /
      ViolationLog / ErrorReport / Medal / UserProfileEntry / TopicSummary / Explanation /
      ChatSession / ChatRawMessage。⚠️ 这些 schema **尚未挂到任何 path**，接口由各板块按里程碑补齐；
      字段名以本节为准，各板块不得另起。
    - **题本升格（D48）**：新增 `ErrorCause` / `ErrorIntent` 两个正交维度，`ErrorRecord` 系列扩展
      `errorCause` / `intent` / `sourceExamId` 三个可选字段（`errorType` 保留为自由文本，向后兼容）。
    - **目标（D6 / D49）**：`Goal` / `GoalSummary` / `GoalCreate` / `GoalUpdate` 增补
      `parentGoalId`（父子树）、`examId` + `targetScore`（引用考试）。`status` 的 enum 未改动。
...

...
servers:
...
security:
  - sessionCookie: []
...
paths:
...
  /timer-sessions:
    post:
      tags: [专注计时]
      operationId: startTimerSession
      summary: 开始计时
      description: |
        已有进行中会话时返回 **409**（不静默接管、也不静默丢弃）：
        专注态同时只能有一个，用户必须先对上一个会话做出处置（继续 / 收尾 / 裁决丢弃），
        否则会出现两段并行计时，或者学习时长被系统悄悄丢掉。

        `countup` 不需要目标时长（传了也会被忽略，不报错——前端切换模式时不该吃 400）。
        `subject` 不传时由服务端从 `taskId` 对应的计划任务推导，推不出才兜底 `other`。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimerSessionStart'
            example:
              mode: countdown
              targetMinutes: 30
              planId: p_1a2b
              taskId: t_3c4d
      responses:
        '201':
          description: 已开始
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerSession'
              example:
                sessionId: ts_9f8e
                mode: countdown
                startedAt: '2026-09-25T19:00:00+08:00'
                targetMinutes: 30
                planId: p_1a2b
                taskId: t_3c4d
                subject: SX
                status: running
                endedAt: null
                effectiveSeconds: null
                lastHeartbeatAt: '2026-09-25T19:00:00+08:00'
                segments:
                  - segmentId: seg_01
                    taskId: t_3c4d
                    startedAt: '2026-09-25T19:00:00+08:00'
                    endedAt: null
                    seconds: null
                createdAt: '2026-09-25T19:00:00+08:00'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '409':
          $ref: '#/components/responses/StateConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /timer-sessions/current:
    get:
      tags: [专注计时]
      operationId: getCurrentTimerSession
      summary: 当前计时会话（恢复视图）
      description: |
        前端进计时页 / 刷新 / 换设备回来时调用。`active=false` 是**正常流程**
        （用户没在计时），不是错误，所以不用 404 表达。

        `restore.needsVerdict=true` 时必须先弹**恢复裁决卡**（D31），**不得自动记账**：
        识别到异常会话（断连超过 30 分钟 / 正计时超 6 小时 / 倒计时到点后又拖超 30 分钟）。
      responses:
        '200':
          description: 当前会话（无会话时 active=false）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerCurrent'
              example:
                active: true
                session:
                  sessionId: ts_9f8e
                  mode: countdown
                  startedAt: '2026-09-25T19:00:00+08:00'
                  targetMinutes: 30
                  planId: p_1a2b
                  taskId: t_3c4d
                  subject: SX
                  status: running
                  endedAt: null
                  effectiveSeconds: null
                  lastHeartbeatAt: '2026-09-25T19:12:00+08:00'
                  segments: []
                  createdAt: '2026-09-25T19:00:00+08:00'
                restore:
                  sessionId: ts_9f8e
                  mode: countdown
                  remainingSeconds: 1080
                  elapsedSeconds: null
                  needsVerdict: false
                  suggestedMinutes: 30
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /timer-sessions/{sessionId}:
    get:
      tags: [专注计时]
      operationId: getTimerSession
      summary: 计时会话详情
      parameters:
        - $ref: '#/components/parameters/TimerSessionId'
      responses:
        '200':
          description: 会话对象（含分段）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerSession'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [专注计时]
      operationId: discardTimerSession
      summary: 丢弃会话（裁决卡「丢弃」）
      description: |
        会话标记为 `abandoned`，**不产生任何学习记录**——这是 D31 验收里
        「僵尸会话不产生记录」的落点：宁可少记一次，也不要凭空造一段没发生过的学习时长。
      parameters:
        - $ref: '#/components/parameters/TimerSessionId'
      responses:
        '200':
          description: 已丢弃
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerDiscarded'
              example:
                discarded: true
                sessionId: ts_9f8e
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /timer-sessions/{sessionId}/heartbeat:
    post:
      tags: [专注计时]
      operationId: heartbeatTimerSession
      summary: 上报心跳
      description: |
        刷新 `lastHeartbeatAt` —— 僵尸判定与「异常会话」识别的唯一依据。
        前端应定时上报（并在每次用户交互时顺带上报）。不刷新心跳的后果不是报错，
        而是**下一次回来会被判成异常会话、要求用户裁决**，所以宁可多报。
      parameters:
        - $ref: '#/components/parameters/TimerSessionId'
      responses:
        '200':
          description: 刷新后的会话
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerSession'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /timer-sessions/{sessionId}/segments:
    post:
      tags: [专注计时]
      operationId: switchTimerTask
      summary: 切换任务（开新的一段）
      description: |
        #14：一个会话内**分段**，每段标注所属任务；**不允许并行计时**，所以只能是"换段"。
        旧段在切换瞬间关闭并结算秒数，新段从此刻开始；结束仍只走一次收尾。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TimerSessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimerSegmentStart'
            example:
              taskId: t_5e6f
      responses:
        '201':
          description: 新的一段
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimerSegment'
              example:
                segmentId: seg_02
                taskId: t_5e6f
                startedAt: '2026-09-25T19:20:00+08:00'
                endedAt: null
                seconds: null
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /timer-sessions/{sessionId}/finish:
    post:
      tags: [专注计时]
      operationId: finishTimerSession
      summary: 结束计时并收尾（落学习记录）
      description: |
        结束会话 + 落学习记录，**一次原子动作**。不在前端分两步（先结束会话、再
        POST /learning-records）：两步之间失败会留下"会话结束了但没有记录"的黑洞，
        正是 D31 要消灭的僵尸形态。

        **有效时长口径（D31：有效时长 ≠ 墙上时长）**：
        - 传了 `durationMinutes` → 以它为准（裁决卡第三态「手动改时长」）；
        - `countdown` → `min(经过, target)`（到点后继续学也不多记，§3.6 封顶）；
        - `countup` → 实际经过；若已判异常且未给时长 → **400 要求先裁决**（不自动记账）。

        响应与 `POST /learning-records` 同构（含同步重算的状态与建议句柄）。
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/TimerSessionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TimerFinish'
            example:
              completion: completed
              selfReport:
                focus: 4
                fatigue: 2
                emotion: positive
                difficultyFeel: moderate
              note: 函数图像终于看懂了
      responses:
        '201':
          description: 学习记录已落库（含状态快照与建议句柄）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningRecordCreated'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/StateConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
      description: |
        HttpOnly Session Cookie（名为 sid），登录成功后由服务端 Set-Cookie 下发，
        有效期 7 天（见 mock-server SESSION_TTL_MS）。浏览器自动携带，前端无需手动设置。
        鉴权失效时业务接口返回 401 UNAUTHENTICATED。
        （历史说明：早期契约曾假设上游 Bearer token 且不含登录/注册接口，
        实际项目改为自带 /auth/* 会话体系，本契约已据此收录，见「鉴权会话」标签。）
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: 可选幂等键，24 小时内重复键直接返回首次结果（用于弱网下重复提交学习记录）
      schema:
        type: string
...
    TimerSessionId:
      name: sessionId
      in: path
      required: true
      description: 计时会话 ID
      schema:
        type: string
...
  responses:
    ValidationFailed:
      description: 参数缺失或取值越界
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: VALIDATION_FAILED
              message: 自评专注度必须为 1-5 的整数
              field: selfReport.focus
    Unauthenticated:
      description: token 缺失或过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHENTICATED
              message: 登录状态已失效，请重新登录
...
    StateConflict:
      description: 当前资源状态不允许该操作（同日计划已存在 / 会话已在运行 / 被引用不可删）
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: STATE_CONFLICT
              message: 当日已存在计划，如需覆盖请传 regenerate=true
    NotFound:
      description: 资源不存在或不属于当前用户
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: 资源不存在
...
    InternalServerError:
      description: 服务端错误。对齐 api-design-unified.md 0.2 节错误码表中的 500 INTERNAL_ERROR。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL_ERROR
              message: 服务开小差了，请稍后再试
...
  schemas:
...
    Subject:
      type: string
      description: 学科代码（拼音首字母大写，与知识点库标准对齐；other 保留兜底）。YW 语文 / SX 数学 / YY 英语 / WL 物理 / HX 化学 / SW 生物 / ZZ 思想政治 / LS 历史 / DL 地理
      enum: [YW, SX, YY, WL, HX, SW, ZZ, LS, DL, other]
...
    Completion:
      type: string
      description: 任务完成度（无 pending）
      enum: [completed, partial, abandoned]
...
    Error:
      type: object
      description: 统一错误格式
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: 错误码
            message:
              type: string
              description: 可展示的错误说明
            field:
              type: string
              description: 校验失败的字段路径，仅参数校验错误时出现
...
    LearningRecordCreated:
      allOf:
        - $ref: '#/components/schemas/LearningRecord'
        - type: object
          description: 提交成功后附带的状态快照与建议句柄
          required: [assessment]
          properties:
            assessment:
              $ref: '#/components/schemas/AssessmentSnapshot'
            recommendation:
              type: object
              nullable: true
              description: 自动创建的建议生成任务；skipRecommendation=true 时为 null
              required: [recommendationId, status]
              properties:
                recommendationId:
                  type: string
                  description: 建议 ID，用于轮询 GET /recommendations/{id}
                status:
                  type: string
                  description: 生成状态；此句柄只出现在刚创建时，取值恒为 pending
                  enum: [pending]
...
    TimerMode:
      type: string
      description: 计时模式。countdown=倒计时（有目标时长）；countup=正计时（不限时）
      enum: [countdown, countup]
    TimerStatus:
      type: string
      description: >-
        会话状态。running=进行中；finished=正常收尾（产生学习记录）；
        abandoned=僵尸裁决后丢弃（**不产生学习记录**）。
      enum: [running, finished, abandoned]
    TimerSession:
      type: object
      description: >-
        计时会话。**服务端持久化「进行中的会话」**（started_at + mode + target_minutes + 计划引用），
        前端只是视图——回来按**真实经过时间**恢复，根治现状「靠路由 state 传参、刷新即丢上下文」。
        只留「结束」一个出口、不做挂起（避免僵尸计时）。
      required: [sessionId, mode, startedAt, status, createdAt]
      properties:
        sessionId: { type: string }
        mode:
          $ref: '#/components/schemas/TimerMode'
        startedAt:
          type: string
          format: date-time
          description: 两种模式都有；恢复计算的基准
        targetMinutes:
          type: integer
          nullable: true
          description: 仅 countdown 有；countup 为 null
        planId: { type: string, nullable: true }
        taskId: { type: string, nullable: true }
        subject:
          $ref: '#/components/schemas/Subject'
        status:
          $ref: '#/components/schemas/TimerStatus'
        endedAt: { type: string, format: date-time, nullable: true }
        effectiveSeconds:
          type: integer
          nullable: true
          description: >-
            **有效时长**（秒），≠ 墙上时长（D31）：倒计时 = min(经过, target)；正计时异常时
            必须用户核实后才写入。不要把 endedAt − startedAt 直接当学习时长。
        lastHeartbeatAt:
          type: string
          format: date-time
          nullable: true
          description: 最近心跳；僵尸判定与「异常会话」识别的依据
        segments:
          type: array
          description: 分段（#14）：切换任务即新开一段，结束只走一次收尾
          items:
            $ref: '#/components/schemas/TimerSegment'
        createdAt: { type: string, format: date-time }
    # ---------- 考试（D49）与计时（D30/D31）的请求体 / 包装 ----------
...
    TimerSessionStart:
      type: object
      description: |
        开始计时的请求体。`targetMinutes` 仅 countdown 必填；
        countup 传了会被忽略（不报错，避免前端切换模式时多余的 400）。
        `subject` 可不传——服务端从 `taskId` 对应的计划任务推导，推不出才兜底 other。
      required: [mode]
      properties:
        mode:
          $ref: '#/components/schemas/TimerMode'
        targetMinutes:
          type: integer
          nullable: true
          minimum: 1
          maximum: 600
        planId:
          type: string
          nullable: true
        taskId:
          type: string
          nullable: true
        subject:
          $ref: '#/components/schemas/Subject'
    TimerSegmentStart:
      type: object
      description: 切换任务请求体（#14：一会话内分段，不允许并行计时）
      properties:
        taskId:
          type: string
          nullable: true
          description: 本段要做什么任务；不传表示自由学习
    TimerCurrent:
      type: object
      description: |
        `GET /timer-sessions/current` 的响应。用 `active` 包装而不是"没有就 404"：
        前端每次进计时页都会问一次，**"没有会话"是正常流程而不是错误**。
      required: [active]
      properties:
        active:
          type: boolean
        session:
          allOf:
            - $ref: '#/components/schemas/TimerSession'
          nullable: true
        restore:
          allOf:
            - $ref: '#/components/schemas/TimerRestore'
          nullable: true
    TimerFinish:
      type: object
      description: |
        计时收尾请求体（D20 三层形态：结束瞬间 0 步 → 轻收尾卡 → 正确率延后补）。
        必填只有 `completion`（唯一"半强制"问句）；`selfReport` **可整段省略**，
        其中专注/疲劳/情绪/难度任一可缺（软字段转译不出就留空，D34 不造数）。
        `durationMinutes` 显式传入时覆盖服务端按 mode 算出的有效时长——
        这正是裁决卡第三态「手动改时长」的落点。
      required: [completion]
      properties:
        completion:
          $ref: '#/components/schemas/Completion'
        selfReport:
          allOf:
            - $ref: '#/components/schemas/RecordSelfReport'
          nullable: true
        durationMinutes:
          type: integer
          nullable: true
          minimum: 1
          maximum: 600
          description: 不传则按 mode 算有效时长；传入即以它为准
        note:
          type: string
          nullable: true
          maxLength: 100
        skipRecommendation:
          type: boolean
          default: false
    TimerDiscarded:
      type: object
      description: 裁决卡「丢弃」的结果。**不产生学习记录**（D31）。
      required: [discarded, sessionId]
      properties:
        discarded:
          type: boolean
        sessionId:
          type: string
...
    TimerSegment:
      type: object
      description: 计时会话内的一段（#14 分段计时）。一个会话可挂多段，每段标注所属任务。
      required: [segmentId, startedAt]
      properties:
        segmentId: { type: string }
        taskId:
          type: string
          nullable: true
          description: 本段属于哪个任务；切换任务即新开一段
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time, nullable: true }
        seconds:
          type: integer
          nullable: true
          description: 本段有效秒数（口径同 effectiveSeconds）
    TimerRestore:
      type: object
      description: >-
        恢复视图（D30 按 mode 分支计算）。`needsVerdict=true` 表示识别到**异常会话**，
        前端必须弹**恢复裁决卡**（D31：「上次计时似乎没正常结束（已计 X 分钟）·
        [保留按 X 记] [丢弃] [手动改时长]」），**不得自动记账**。
      required: [sessionId, mode, needsVerdict]
      properties:
        sessionId: { type: string }
        mode:
          $ref: '#/components/schemas/TimerMode'
        remainingSeconds:
          type: integer
          nullable: true
          description: 仅 countdown：剩余 = target − 已过；≤0 视为已到点（软提醒，不自动结束）
        elapsedSeconds:
          type: integer
          nullable: true
          description: 仅 countup：累计 = now − startedAt
        needsVerdict:
          type: boolean
          description: 是否需用户裁决（正计时超上限 / 倒计时到点后长时间未收尾）
        suggestedMinutes:
          type: integer
          nullable: true
          description: 裁决卡「保留按 X 记」的预填值（倒计时封顶 target；正计时按截断点）

    # ---------- G 板块：埋点事件（D39 / §8） ----------
...